--- CLAUDE.md
@@ -1,2 +1,0 @@
-# Agent Guidelines
-AI Agents must read `docs/DEV_CONVENTION.md` before starting any development.
